[DUT-874]: packages/domain 초안 분리
