+++ apps/app/src/entities/shift/model/request.ts
@@ -1,14 +1,1 @@
-export type TDutyRequest = {
-    wardReqShiftId: number;
-    nurseId: number;
-    nurseName: string;
-    /**신청근무 날짜 */
-    date: number;
-    /**신청근무를 신청한 날짜 */
-    requestDate: string;
-    wardShiftTypeId: number;
-    wardShiftTypeShortName: string;
-    wardShiftTypeColor: string;
-    isRead: boolean;
-    isAccepted: boolean | null;
-};
+export type {TDutyRequest} from '@dutying/domain/shift';
