+++ packages/domain/src/index.ts
@@ -1,0 +1,4 @@
+export type * from './account';
+export type * from './nurse';
+export type * from './shift';
+export type * from './ward';
